feat(qmi): add QRTR transport support and modularize QMI protocol implementation
This commit introduces a modular architecture for QMI communication with
separate transport layers and protocol abstractions. Key changes include:

- Introduce QRTR protocol support alongside existing QMI transport
- Create dedicated core module with protocol definitions and constants
- Implement transport abstraction layer for multiple communication methods
- Add comprehensive QMI error definitions and type-safe TLV operations
- Migrate MBIM transport methods to use int64 return types for consistency

The new architecture enables flexible transport mechanisms while
maintaining clean separation between protocol logic and communication layers.

+++ driver/mbim/types.go
@@ -20,19 +20,19 @@ type Request struct {
 	Response      encoding.BinaryUnmarshaler
 }
 
-func (r *Request) WriteTo(w net.Conn) (int, error) {
+func (r *Request) WriteTo(w net.Conn) (int64, error) {
 	data, err := r.MarshalBinary()
 	if err != nil {
 		return 0, err
 	}
 	n, err := w.Write(data)
 	if err != nil {
-		return n, err
+		return int64(n), err
 	}
-	return n, nil
+	return int64(n), nil
 }
 
-func (r *Request) ReadFrom(c net.Conn) (int, error) {
+func (r *Request) ReadFrom(c net.Conn) (int64, error) {
 	if r.ReadTimeout == 0 {
 		r.ReadTimeout = 30 * time.Second
 	}
@@ -65,7 +65,7 @@ func (r *Request) ReadFrom(c net.Conn) (int, error) {
 		if err := response.UnmarshalBinary(buf); err != nil {
 			return 0, err
 		}
-		return len(buf), nil
+		return int64(len(buf)), nil
 	}
 	return 0, fmt.Errorf("transaction ID %d not found in response", r.TransactionID)
 }
